requirements.txt
airtest==1.2.3
airtest-selenium==1.0.3
amqp==2.6.1
appnope==0.1.2
asgiref==3.4.1
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Deprecated==1.2.13
et-xmlfile==1.1.0
facebook-wda==1.4.3
Faker==10.0.0
hrpc==1.0.9
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.2
jsonpath==0.82
kombu==4.6.11
lemon-easytest==1.0a3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mss==4.0.3
mysqlclient==2.0.3
numpy==1.19.3
opencv-contrib-python==4.5.4.58
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
pocoui==1.0.85
prompt-toolkit==3.0.23
ptyprocess==0.7.0
py==1.11.0
Pygments==2.10.0
pynput==1.7.6
pyparsing==2.4.7
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-xlib==0.31
pytz==2021.3
pywinauto==0.6.3
PyYAML==6.0
redis==4.1.0
requests==2.26.0
retry==0.9.2
ruamel.yaml==0.16.7
selenium==3.141.0
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.2
style==1.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.1
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.48.0
wrapt==1.13.3
